requirements.txt
landingai-ade==0.17.1
python-dotenv==1.1.1
loguru==0.7.3
PyMuPDF==1.26.5
pillow==11.3.0
requests==2.32.5
openai==2.3.0
pandas==2.3.3
tabulate==0.9.0
scikit-learn==1.7.2
opencv-python==4.12.0.88
pytesseract==0.3.13
pydantic_settings==2.11.0